flask-server/requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
distlib==0.3.8
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
pipenv==2024.0.1
platformdirs==4.2.2
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
referencing==0.35.1
rpds-py==0.19.0
setuptools==71.0.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
virtualenv==20.26.3
Werkzeug==3.0.3
